// src/app/faturamento/parametrofaturamento/models/parametrofaturamento.ts
export class ParametroFaturamento{
    id: number;
    bloqueioAcesso: string;
    fechamentoMes: string;
    emiteAvisosBoletosVencidos: string;
    emiteAvisosReajustes: string;
    quantidadeParcelasEmAberto: number;
    logoEmpresa: string;
    gerarComFaturasVencidas: string;
    calcularDescontosAdimplentes: string;
    considerarDescontoPara: string;
    quantidadeMesesDesconto: number;
    percentual: number;
    faturarInativos: string;
    faturarClientesConstituicao: string;
    calcularJuros: string;
    percentualJuros: number;
    calcularMulta: string;
    percentualMulta: number;
    dataVencimentoDefault: string;
    calcularJurosRenegociacao: string;
    percentualJurosRenegociacao: number;
    calcularMultaRenegociacao: string;
    percentualMultaRenegociacao: number;
    valorMinimoIRRF: number;
    fatoGeradorIRRF: string;
    valorMinimoINSSRetido: number;
    valorMinimoISSRetido: number;
    crf: string;
    aliquotaCRF: number;
    pis: string;
    aliquotaPIS: number;
    cofins: string;
    aliquotaCOFINS: number;
    csll: string;
    aliquotaCSLL: number;
    inss: string;
    aliquotaINSS: number;
    iss: string;
    aliquotaISS: number;
    valorMinimoCalculo: number;
    fatoGeradorCRFPISCOFINS: string;
    deduzirIRRFCRFValorReceber: string;
    notaFiscal: string;
    serieRPS: string;
    servicoId: number;
    gerarValorAproxImpostosIBPT: string;
    aliquotaRetencaoFonte: number;
    aliquotaIBPTAte2014: number;
    aliquotaIBPTAPartir2015: number;
    aliquotaIBPTMunicipal: number;
    descricao: string;
    empresaId: number;
    grupoEmpresaId: number;
    configuracaoPagamentoId: number; 
    contaCorrenteId: number; 
    valorMinimoCRF: number;
    excluido: string;
};

export class ConfiguracaoPagamento {
    id: string;
    descricao: string;
};

export class ContaCorrente {
    id: string;
    descricao: string;
};